Programacion-web-master/requirements.txt
Django==5.0.2
stripe==8.4.0
requests==2.31.0
Pillow==10.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
cx_Oracle==8.3.0
django-cors-headers==4.3.0 